backend/requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
Deprecated==1.2.14
Flask==2.3.2
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
urllib3==2.0.4
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
